requirements.txt
groq
python-dotenv
colorama
flask
flask_cors